+++ apps/mobile/src/domain/user/User.ts
