+++ apps/mobile/src/infrastructure/service.ts
@@ -1,4 +1,4 @@
-import { UserService } from '../domain/post/UserService';
+import { UserService } from '../domain/user/UserService';
 import { TelepactHttpService } from './TelepactHttpService';
 
 export const telepactService = new TelepactHttpService("http://192.168.0.15:8082");
